+++ docs/components/Home/Hero.vue
@@ -1,113 +1,60 @@
 <template>
-  <div class="">
-    <div class="relative isolate px-6 pt-14 lg:px-8">
-      <div
-        class="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80"
-        aria-hidden="true"
-      >
-        <div
-          class="relative left-[calc(50%-11rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#eab308] to-[#f97316] opacity-20 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
-          style="
-            clip-path: polygon(
-              74.1% 44.1%,
-              100% 61.6%,
-              97.5% 26.9%,
-              85.5% 0.1%,
-              80.7% 2%,
-              72.5% 32.5%,
-              60.2% 62.4%,
-              52.4% 68.1%,
-              47.5% 58.3%,
-              45.2% 34.5%,
-              27.5% 76.7%,
-              0.1% 64.9%,
-              17.9% 100%,
-              27.6% 76.8%,
-              76.1% 97.7%,
-              74.1% 44.1%
-            );
-          "
-        ></div>
-      </div>
+  <div class="bg-surface-primary dark:bg-surface-primary-dark relative overflow-hidden">
+    <HeroLines />
 
-      <div class="mx-auto max-w-2xl py-32 sm:py-48 lg:py-56">
-        <div class="hidden sm:mb-8 sm:flex sm:justify-center">
-          <div
-            class="bg-white dark:bg-slate-700 relative rounded-full px-3 py-1 text-sm leading-6 text-gray-600 dark:text-gray-300 ring-1 ring-gray-900/10 hover:ring-gray-900/20"
+    <div class="relative pt-40 pb-16 px-4">
+      <div class="mx-auto max-w-4xl text-center">
+        <div class="inline-flex rounded-full p-0.5 rotate-2 bg-gradient-to-tr from-tertiary to-secondary">
+          <h1
+            class="bg-white dark:bg-slate-800 rounded-full px-3 py-1 text-sm leading-6 text-primary dark:text-text-secondary-dark backdrop-blur-sm"
           >
-            Stimulus Components is now an organization on NPM.
-
-            <a href="https://www.npmjs.com/org/stimulus-components" class="font-semibold text-orange-500">
-              <span class="absolute inset-0" aria-hidden="true"></span>
-              Read more
-              <span aria-hidden="true">&rarr;</span>
-            </a>
-          </div>
+            Stimulus Components
+          </h1>
         </div>
 
-        <div class="text-center">
-          <h1 class="text-4xl tracking-tight font-extrabold text-gray-900 dark:text-white sm:text-5xl md:text-6xl">
-            <span class="block xl:inline">Stimulus</span>
-            <span class="block bg-gradient-to-r from-yellow-500 to-orange-500 xl:inline bg-clip-text text-transparent">
-              Components</span
-            >
-          </h1>
+        <div class="mt-8 mb-20 text-balance">
+          <h2 class="text-5xl font-extrabold leading-16 text-primary dark:text-primary-dark">
+            Craft world-class Stimulus controllers with your own styles
+          </h2>
 
-          <p class="mt-6 text-lg leading-8 text-gray-600 dark:text-gray-400">
-            Stimulus Components is an open-source set of StimulusJS controllers to solve common patterns.
+          <p class="text-text-secondary dark:text-text-secondary-dark font-light leading-8 mt-3">
+            Over 25 Stimulus controllers with built-in behavior, top-tier documentation, designed for easy extension,
+            and ready for your styles to solve every day frontend problems.
           </p>
+        </div>
+      </div>
 
-          <div class="mt-10 sm:flex sm:justify-center">
-            <div class="rounded-md shadow">
-              <NuxtLink
-                to="/docs"
-                class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white md:py-4 md:text-lg md:px-10 bg-gradient-to-r from-yellow-500 to-orange-500"
-              >
-                Get started &rarr;
-              </NuxtLink>
-            </div>
-
-            <div class="mt-3 rounded-md shadow sm:mt-0 sm:ml-3">
-              <a
-                href="https://github.com/stimulus-components"
-                class="w-full flex items-center justify-center gap-1 px-8 py-3 border border-transparent text-base font-medium rounded-md bg-white hover:bg-gray-50 md:py-4 md:text-lg md:px-10"
-              >
-                <GithubIcon />
-
-                GitHub
-              </a>
-            </div>
+      <div class="max-w-6xl mx-auto pt-8 grid grid-cols-1 md:grid-cols-3 gap-4">
+        <div
+          v-for="(component, index) in components"
+          :key="component.title"
+          :class="[
+            'p-0.5 rounded-3xl bg-gradient-to-tr from-tertiary to-secondary',
+            {
+              'md:translate-x-1/3': index === 0,
+              'md:-translate-y-1/4 z-40': index === 1,
+              'md:-translate-x-1/3 z-30': index === 2,
+              'z-10': index === 3,
+              'md:-rotate-6': index === 0 || index === 3,
+              'md:rotate-6': index === 2 || index === 4,
+              'md:-translate-y-1/4 md:translate-x-1/2': index >= 3,
+            },
+          ]"
+        >
+          <div class="bg-gray-50 dark:bg-slate-800 relative rounded-3xl h-full">
+            <component :is="componentsDemo[component.package]" :is-home="true" />
           </div>
         </div>
       </div>
 
-      <div
-        class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]"
-        aria-hidden="true"
-      >
-        <div
-          class="relative left-[calc(50%+3rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 bg-gradient-to-tr from-[#eab308] to-[#f97316] opacity-20 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
-          style="
-            clip-path: polygon(
-              74.1% 44.1%,
-              100% 61.6%,
-              97.5% 26.9%,
-              85.5% 0.1%,
-              80.7% 2%,
-              72.5% 32.5%,
-              60.2% 62.4%,
-              52.4% 68.1%,
-              47.5% 58.3%,
-              45.2% 34.5%,
-              27.5% 76.7%,
-              0.1% 64.9%,
-              17.9% 100%,
-              27.6% 76.8%,
-              76.1% 97.7%,
-              74.1% 44.1%
-            );
-          "
-        ></div>
+      <div class="py-12 flex flex-col sm:flex-row items-center justify-center gap-4">
+        <UIButton class="block" variant="tertiary" href="/docs">Get started</UIButton>
+
+        <UIButton class="block" href="https://github.com/stimulus-components/stimulus-components" target="_blank">
+          <GithubIcon />
+
+          GitHub
+        </UIButton>
       </div>
     </div>
   </div>
@@ -115,4 +62,29 @@
 
 <script setup>
 import GithubIcon from "@/components/Icons/GithubIcon.vue"
+import CharacterCounter from "@/components/content/Demo/CharacterCounter.vue"
+import CheckboxSelectAll from "@/components/content/Demo/CheckboxSelectAll.vue"
+import Clipboard from "@/components/content/Demo/Clipboard.vue"
+import Dropdown from "@/components/content/Demo/Dropdown.vue"
+import PasswordVisibility from "@/components/content/Demo/PasswordVisibility.vue"
+import HeroLines from "@/components/Home/HeroLines.vue"
+
+const { data: components } = await useAsyncData("components-hero", () =>
+  queryContent("docs")
+    .sort({ title: 1 })
+    .where({
+      package: {
+        $in: ["character-counter", "checkbox-select-all", "clipboard", "dropdown", "password-visibility"],
+      },
+    })
+    .find(),
+)
+
+const componentsDemo = {
+  "character-counter": CharacterCounter,
+  "checkbox-select-all": CheckboxSelectAll,
+  clipboard: Clipboard,
+  dropdown: Dropdown,
+  "password-visibility": PasswordVisibility,
+}
 </script>