+++ docs/components/Home/Testimonials.vue
@@ -1,95 +1,101 @@
 <template>
-  <div class="py-24 sm:py-32">
-    <div class="mx-auto max-w-7xl px-6 lg:px-8">
-      <div class="mx-auto max-w-xl text-center">
-        <h2 class="text-lg font-semibold leading-8 tracking-tight text-orange-600">Testimonials</h2>
-        <p class="mt-2 text-3xl font-bold tracking-tight text-gray-900 dark:text-white sm:text-4xl">
-          See what other developers are saying about it!
-        </p>
-      </div>
+  <div class="pt-16 mx-auto max-w-6xl px-4 text-center">
+    <div class="inline-flex rounded-full p-0.5 -rotate-2 bg-gradient-to-tr from-surface-primary to-secondary">
+      <h2
+        class="bg-white dark:bg-slate-800 relative rounded-full px-3 py-1 text-sm leading-6 text-primary dark:text-text-secondary-dark backdrop-blur-sm"
+      >
+        Testimonials
+      </h2>
+    </div>
 
-      <div class="mx-auto mt-16 flow-root max-w-2xl sm:mt-20 lg:mx-0 lg:max-w-none">
-        <div class="-mt-8 sm:-mx-4 sm:columns-2 lg:columns-3">
-          <div
-            v-for="testimonial in testimonials"
-            :key="testimonial.author"
-            class="pt-8 sm:inline-block sm:w-full sm:px-4"
-          >
-            <figure class="rounded-2xl bg-gray-50 dark:bg-slate-900 p-8 text-sm leading-6">
-              <blockquote class="text-gray-900 dark:text-gray-400">
-                <p>“{{ testimonial.content }}”</p>
-              </blockquote>
+    <p class="text-4xl font-bold leading-8 tracking-tight text-primary dark:text-primary-dark mt-2">
+      See what other developers are saying about it!
+    </p>
 
-              <figcaption>
-                <a
-                  :href="`https://twitter.com/${testimonial.twitter}`"
-                  class="mt-6 flex items-center gap-x-4"
-                  target="_blank"
-                >
-                  <img
-                    :src="`/testimonials/${testimonial.twitter}.jpg`"
-                    :alt="testimonial.name"
-                    class="size-10 rounded-full bg-gray-50"
-                    loading="lazy"
-                  />
-
-                  <div>
-                    <div class="font-semibold text-gray-900 dark:text-white">{{ testimonial.author }}</div>
-
-                    <div class="text-gray-600 dark:text-gray-400">{{ testimonial.job }}</div>
-                  </div>
-                </a>
-              </figcaption>
-            </figure>
-          </div>
-        </div>
+    <div class="text-left mt-8">
+      <div class="sm:-mx-4 sm:columns-2 lg:columns-3">
+        <Testimonial
+          v-for="(testimonial, index) in testimonials"
+          :key="testimonial.author"
+          :testimonial="testimonial"
+          :index="index"
+        />
       </div>
     </div>
   </div>
 </template>
 
 <script setup>
+import Testimonial from "@/components/Home/Testimonial.vue"
+
 const testimonials = [
   {
     author: "Guillaume Briday",
     content:
       "All controllers originate from real-world applications that addressed common problems faced by nearly everyone.",
-    twitter: "guillaumebriday",
+    username: "guillaumebriday",
     job: "Creator of Stimulus Components",
   },
   {
     author: "Matt Polito",
     content:
       "They’re lightweight, can be pulled in individually and solve a problem… it would be a disservice to our clients NOT to use them.",
-    twitter: "mattpolito",
+    username: "mattpolito",
     job: "Senior Engineer at Hashrocket",
   },
   {
     author: "Avi Flombaum",
     content:
       "Stimulus Components are indispensable libraries for common UI patterns that are expertly developed and maintained.",
-    twitter: "aviflombaum",
+    username: "aviflombaum",
     job: "Co-founder of Flatiron School",
   },
   {
     author: "Ryan Weaver",
     content: "High-quality, pre-made stimulus controllers that solve my most common problems? I’m in love!",
-    twitter: "weaverryan",
+    username: "weaverryan",
     job: "Writer for SymfonyCasts & Symfony core member",
   },
   {
     author: "Mike Holford",
     content:
       "Just used my first component today for https://urbanhunt.co 🎉 Super easy to integrate and got nested forms working in 10 minutes. Thanks!",
-    twitter: "holfyio",
+    username: "holfyio",
     job: "Developer / Maker",
   },
   {
     author: "Adrian Marin",
     content:
       "From time to time I need to apply a common pattern that I haven't yet built with Stimulus. Stimulus Components has it already! The components are polished, easy to implement, and the docs are great!",
-    twitter: "adrianthedev",
+    username: "adrianthedev",
     job: "Creator of @avo_hq",
+  },
+  {
+    author: "Adrian Siami",
+    content:
+      "Inspiring and smart stimulus controllers that will get you a head start, a great toolbox for any sensible developer!",
+    username: "intrepidd",
+    job: "Open source contributor",
+  },
+  {
+    author: "Laurent Buffevant",
+    content:
+      "Stimulus Components has become an essential tool in our daily workflow. Its ease of use, adaptability, and pre-built features save us time while enabling us to deliver outstanding interactivity for our clients.",
+    username: "studiohb",
+    job: "Co-founder of Studio HB",
+  },
+  {
+    author: "Cyril Blaecke",
+    content:
+      "Why reinvent the wheel? Stimulus Components are time saving and easy to integrate. We use them in almost all of our Rails projects at Dilolabs.",
+    username: "cbldev",
+    job: "Founder of Dilolabs",
+  },
+  {
+    author: "Yaroslav Shmarov",
+    content: "Don't reinvent the wheel. Use Stimulus Components. I use them in all my projects.",
+    username: "yarotheslav",
+    job: "Founder of SupeRails.com",
   },
 ]
 </script>