+++ docs/components/Icons/NpmIcon.vue
@@ -1,0 +1,14 @@
+<template>
+  <svg
+    fill="currentColor"
+    focusable="false"
+    height="1em"
+    stroke="currentColor"
+    stroke-width="0"
+    viewBox="0 0 16 16"
+    width="1em"
+    class="text-[#E53E3E]"
+  >
+    <path d="M0 0v16h16v-16h-16zM13 13h-2v-8h-3v8h-5v-10h10v10z"></path>
+  </svg>
+</template>