+++ docs/components/Icons/SimpleIcon.vue
@@ -1,0 +1,12 @@
+<template>
+  <svg width="48" height="48" fill="none" aria-hidden="true">
+    <path
+      d="M5 11a4 4 0 0 1 4-4h30a4 4 0 0 1 4 4v26a4 4 0 0 1-4 4H9a4 4 0 0 1-4-4V11Z"
+      fill="currentColor"
+      fill-opacity=".1"
+      stroke="currentColor"
+      stroke-width="2"
+    ></path>
+    <path d="M15 7v34" stroke="currentColor" stroke-width="2" stroke-linecap="round"></path>
+  </svg>
+</template>