+++ components/checkbox-select-all/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.1.0] - 2024-11-12
+
+### Chore
+
+- Add an option to disable the indeterminate state.
+
 ## [6.0.0] - 2024-02-27
 
 ### Chore